# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NetSuite Record-File Viewer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReloadDataResult {
  allFilesCount: number;
  recordAttachmentsCount: number;
  durationMs: number;
}

export type DataUploadInputType = typeof DataUploadInputType[keyof typeof DataUploadInputType];


export const DataUploadInputType = {
  all_files: 'all_files',
  record_attachments: 'record_attachments',
  deletion_log: 'deletion_log',
} as const;

export interface DataUploadInput {
  type: DataUploadInputType;
  /** Raw file content (CSV or pipe-delimited) */
  content: string;
  /** Optional label for the upload — used for deletion_log to track which logs have been ingested */
  source?: string;
}

export interface CompletionPerType {
  recordType: string;
  totalOrigs: number;
  deletedCount: number;
  remainingCount: number;
  coveragePercent: number;
  errorCount: number;
  protectedCount: number;
}

export interface CompletionOverall {
  totalOrigs: number;
  totalDeleted: number;
  totalRemaining: number;
  totalErrors: number;
  totalProtected: number;
  coveragePercent: number;
}

export type CompletionResponseStatusCounts = {[key: string]: number};

export interface CompletionResponse {
  perType: CompletionPerType[];
  overall: CompletionOverall;
  deletionLogEntryCount: number;
  sources: string[];
  statusCounts: CompletionResponseStatusCounts;
}

export interface CompletionErrorRecord {
  recordType: string;
  recordId: string;
  recordName: string;
}

export interface CompletionErrorEntry {
  fileId: string;
  status: string;
  /** @nullable */
  fileName?: string | null;
  records: CompletionErrorRecord[];
}

export interface CompletionErrorsResponse {
  errors: CompletionErrorEntry[];
  totalErrorFiles: number;
}

export interface DataUploadResult {
  success: boolean;
  rowCount: number;
  type: string;
}

export interface DataStatus {
  allFilesLoaded: boolean;
  recordAttachmentsLoaded: boolean;
  allFilesCount: number;
  recordAttachmentsCount: number;
}

export interface RecordTypeSummary {
  recordType: string;
  recordCount: number;
  fileCount: number;
  stubCount: number;
  missingStubCount: number;
  totalSizeBytes: number;
}

export interface AttachedFile {
  fileId: string;
  fileName: string;
  fileType: string;
  sizeBytes: number;
  hasStub: boolean;
  isStubFile?: boolean;
  recordStatus: string;
  /** @nullable */
  stubFileName?: string | null;
  /** @nullable */
  folderId?: string | null;
  /** @nullable */
  folderName?: string | null;
}

export interface RecordItem {
  recordId: string;
  recordName: string;
  recordType: string;
  recordStatus: string;
  fileCount: number;
  stubCount: number;
  missingStubCount: number;
}

export interface RecordListResponse {
  records: RecordItem[];
  total: number;
}

export interface FileItem {
  fileId: string;
  fileName: string;
  /** @nullable */
  folderId: string | null;
  /** @nullable */
  folderName: string | null;
  fileType: string;
  sizeBytes: number;
  hasStub: boolean;
  attachedRecordCount: number;
}

export interface FileListResponse {
  files: FileItem[];
  total: number;
}

export interface DashboardSummary {
  totalFiles: number;
  totalRecords: number;
  totalRecordTypes: number;
  totalAttachments: number;
  filesWithStub: number;
  filesMissingStub: number;
  stubCoveragePercent: number;
  totalSizeBytes: number;
}

export interface VerificationItem {
  recordType: string;
  recordId: string;
  recordName: string;
  recordStatus: string;
  fileId: string;
  fileName: string;
  sizeBytes: number;
  fileType: string;
}

export interface VerificationResponse {
  items: VerificationItem[];
  total: number;
}

export interface StubCoverageByType {
  recordType: string;
  fileCount: number;
  stubCount: number;
  missingStubCount: number;
  coveragePercent: number;
  totalSizeBytes: number;
}

export interface CompletionStateRow {
  recordType: string;
  manuallyDone: boolean;
  notes: string;
}

export interface CompletionStatePatch {
  manuallyDone?: boolean;
  notes?: string;
}

export interface NetsuiteStatus {
  configured: boolean;
  /** @nullable */
  accountId: string | null;
  /** @nullable */
  message?: string | null;
}

export interface RefreshStubInput {
  recordType: string;
}

export interface RefreshStubResult {
  success: boolean;
  updated: number;
  message: string;
}

export type ListRecordsParams = {
recordType?: string;
search?: string;
stubStatus?: ListRecordsStubStatus;
sort?: ListRecordsSort;
limit?: number;
offset?: number;
};

export type ListRecordsStubStatus = typeof ListRecordsStubStatus[keyof typeof ListRecordsStubStatus];


export const ListRecordsStubStatus = {
  all: 'all',
  has_stub: 'has_stub',
  missing_stub: 'missing_stub',
} as const;

export type ListRecordsSort = typeof ListRecordsSort[keyof typeof ListRecordsSort];


export const ListRecordsSort = {
  missing_stubs_desc: 'missing_stubs_desc',
  missing_stubs_asc: 'missing_stubs_asc',
  record_name: 'record_name',
  record_type: 'record_type',
} as const;

export type PatchCompletionState200 = {
  success: boolean;
};

export type ClearDeletionLog200 = {
  success: boolean;
};

export type ListAllFilesParams = {
folderId?: string;
search?: string;
stubStatus?: ListAllFilesStubStatus;
limit?: number;
offset?: number;
};

export type ListAllFilesStubStatus = typeof ListAllFilesStubStatus[keyof typeof ListAllFilesStubStatus];


export const ListAllFilesStubStatus = {
  all: 'all',
  has_stub: 'has_stub',
  missing_stub: 'missing_stub',
} as const;

export type GetVerificationItemsParams = {
recordType?: string;
limit?: number;
offset?: number;
};