# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NetSuite Record-File Viewer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: data
    description: File and record data operations
  - name: netsuite
    description: NetSuite connectivity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /data/upload:
    post:
      operationId: uploadData
      tags: [data]
      summary: Upload CSV/pipe-delimited data files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataUploadInput"
      responses:
        "200":
          description: Upload accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataUploadResult"

  /data/status:
    get:
      operationId: getDataStatus
      tags: [data]
      summary: Get status of loaded data
      responses:
        "200":
          description: Current data status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataStatus"

  /data/record-types:
    get:
      operationId: listRecordTypes
      tags: [data]
      summary: List all distinct record types
      responses:
        "200":
          description: List of record types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecordTypeSummary"

  /data/records:
    get:
      operationId: listRecords
      tags: [data]
      summary: List records, optionally filtered by type
      parameters:
        - name: recordType
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: stubStatus
          in: query
          required: false
          schema:
            type: string
            enum: [all, has_stub, missing_stub]
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [missing_stubs_desc, missing_stubs_asc, record_name, record_type]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of records with their file attachments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordListResponse"

  /data/completion:
    get:
      operationId: getCompletion
      tags: [data]
      summary: Phase 3 completion stats per record type
      responses:
        "200":
          description: Completion stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionResponse"

  /data/completion-state:
    get:
      operationId: getCompletionState
      tags: [data]
      summary: Get persisted completion state (checkboxes + notes) for all record types
      responses:
        "200":
          description: Array of completion state rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompletionStateRow"

  /data/completion-state/{recordType}:
    patch:
      operationId: patchCompletionState
      tags: [data]
      summary: Upsert manuallyDone and/or notes for a record type
      parameters:
        - name: recordType
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionStatePatch"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean

  /data/completion/errors:
    get:
      operationId: getCompletionErrors
      tags: [data]
      summary: List files with error statuses in the deletion log, with their attached records
      responses:
        "200":
          description: Error entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionErrorsResponse"

  /data/deletion-log:
    delete:
      operationId: clearDeletionLog
      tags: [data]
      summary: Clear all loaded deletion log rows
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean

  /data/records/{recordType}/{recordId}/files:
    get:
      operationId: getRecordFiles
      tags: [data]
      summary: Get all files attached to a specific record
      parameters:
        - name: recordType
          in: path
          required: true
          schema:
            type: string
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Files attached to the record
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttachedFile"
        "404":
          description: Record not found

  /data/files:
    get:
      operationId: listAllFiles
      tags: [data]
      summary: List all files with optional filters
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: stubStatus
          in: query
          required: false
          schema:
            type: string
            enum: [all, has_stub, missing_stub]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"

  /data/verification:
    get:
      operationId: getVerificationItems
      tags: [data]
      summary: Files missing stubs with their record context, sorted by record type
      parameters:
        - name: recordType
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Missing-stub file-record pairs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"

  /data/reload:
    post:
      operationId: reloadData
      tags: [data]
      summary: Reload all data from disk files (truncates DB and re-streams /data/ directory)
      responses:
        "200":
          description: Reload complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReloadDataResult"

  /data/summary:
    get:
      operationId: getSummary
      tags: [data]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /data/stub-coverage:
    get:
      operationId: getStubCoverage
      tags: [data]
      summary: Get stub coverage breakdown by record type
      responses:
        "200":
          description: Stub coverage by record type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StubCoverageByType"

  /netsuite/status:
    get:
      operationId: getNetsuiteStatus
      tags: [netsuite]
      summary: Check if NetSuite credentials are configured
      responses:
        "200":
          description: NetSuite connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetsuiteStatus"

  /netsuite/refresh-stubs:
    post:
      operationId: refreshStubStatus
      tags: [netsuite]
      summary: Refresh stub status from NetSuite for a record type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshStubInput"
      responses:
        "200":
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshStubResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ReloadDataResult:
      type: object
      required: [allFilesCount, recordAttachmentsCount, durationMs]
      properties:
        allFilesCount:
          type: integer
        recordAttachmentsCount:
          type: integer
        durationMs:
          type: integer

    DataUploadInput:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [all_files, record_attachments, deletion_log]
        content:
          type: string
          description: Raw file content (CSV or pipe-delimited)
        source:
          type: string
          description: Optional label for the upload — used for deletion_log to track which logs have been ingested

    CompletionPerType:
      type: object
      required: [recordType, totalOrigs, deletedCount, remainingCount, coveragePercent, errorCount, protectedCount]
      properties:
        recordType: { type: string }
        totalOrigs: { type: integer }
        deletedCount: { type: integer }
        remainingCount: { type: integer }
        coveragePercent: { type: number }
        errorCount: { type: integer }
        protectedCount: { type: integer }

    CompletionOverall:
      type: object
      required: [totalOrigs, totalDeleted, totalRemaining, totalErrors, totalProtected, coveragePercent]
      properties:
        totalOrigs: { type: integer }
        totalDeleted: { type: integer }
        totalRemaining: { type: integer }
        totalErrors: { type: integer }
        totalProtected: { type: integer }
        coveragePercent: { type: number }

    CompletionResponse:
      type: object
      required: [perType, overall, deletionLogEntryCount, sources, statusCounts]
      properties:
        perType:
          type: array
          items: { $ref: "#/components/schemas/CompletionPerType" }
        overall: { $ref: "#/components/schemas/CompletionOverall" }
        deletionLogEntryCount: { type: integer }
        sources:
          type: array
          items: { type: string }
        statusCounts:
          type: object
          additionalProperties: { type: integer }

    CompletionErrorRecord:
      type: object
      required: [recordType, recordId, recordName]
      properties:
        recordType: { type: string }
        recordId: { type: string }
        recordName: { type: string }

    CompletionErrorEntry:
      type: object
      required: [fileId, status, records]
      properties:
        fileId: { type: string }
        status: { type: string }
        fileName: { type: ["string", "null"] }
        records:
          type: array
          items: { $ref: "#/components/schemas/CompletionErrorRecord" }

    CompletionErrorsResponse:
      type: object
      required: [errors, totalErrorFiles]
      properties:
        errors:
          type: array
          items: { $ref: "#/components/schemas/CompletionErrorEntry" }
        totalErrorFiles: { type: integer }

    DataUploadResult:
      type: object
      required: [success, rowCount, type]
      properties:
        success:
          type: boolean
        rowCount:
          type: integer
        type:
          type: string

    DataStatus:
      type: object
      required: [allFilesLoaded, recordAttachmentsLoaded, allFilesCount, recordAttachmentsCount]
      properties:
        allFilesLoaded:
          type: boolean
        recordAttachmentsLoaded:
          type: boolean
        allFilesCount:
          type: integer
        recordAttachmentsCount:
          type: integer

    RecordTypeSummary:
      type: object
      required: [recordType, recordCount, fileCount, stubCount, missingStubCount, totalSizeBytes]
      properties:
        recordType:
          type: string
        recordCount:
          type: integer
        fileCount:
          type: integer
        stubCount:
          type: integer
        missingStubCount:
          type: integer
        totalSizeBytes:
          type: integer

    AttachedFile:
      type: object
      required: [fileId, fileName, fileType, sizeBytes, hasStub, recordStatus]
      properties:
        fileId:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        sizeBytes:
          type: integer
        hasStub:
          type: boolean
        isStubFile:
          type: boolean
        recordStatus:
          type: string
        stubFileName:
          type: ["string", "null"]
        folderId:
          type: ["string", "null"]
        folderName:
          type: ["string", "null"]

    RecordItem:
      type: object
      required: [recordId, recordName, recordType, recordStatus, fileCount, stubCount, missingStubCount]
      properties:
        recordId:
          type: string
        recordName:
          type: string
        recordType:
          type: string
        recordStatus:
          type: string
        fileCount:
          type: integer
        stubCount:
          type: integer
        missingStubCount:
          type: integer

    RecordListResponse:
      type: object
      required: [records, total]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/RecordItem"
        total:
          type: integer

    FileItem:
      type: object
      required: [fileId, fileName, folderId, folderName, fileType, sizeBytes, hasStub, attachedRecordCount]
      properties:
        fileId:
          type: string
        fileName:
          type: string
        folderId:
          type: ["string", "null"]
        folderName:
          type: ["string", "null"]
        fileType:
          type: string
        sizeBytes:
          type: integer
        hasStub:
          type: boolean
        attachedRecordCount:
          type: integer

    FileListResponse:
      type: object
      required: [files, total]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
        total:
          type: integer

    DashboardSummary:
      type: object
      required: [totalFiles, totalRecords, totalRecordTypes, totalAttachments, filesWithStub, filesMissingStub, stubCoveragePercent, totalSizeBytes]
      properties:
        totalFiles:
          type: integer
        totalRecords:
          type: integer
        totalRecordTypes:
          type: integer
        totalAttachments:
          type: integer
        filesWithStub:
          type: integer
        filesMissingStub:
          type: integer
        stubCoveragePercent:
          type: number
        totalSizeBytes:
          type: integer

    VerificationItem:
      type: object
      required: [recordType, recordId, recordName, recordStatus, fileId, fileName, sizeBytes, fileType]
      properties:
        recordType:
          type: string
        recordId:
          type: string
        recordName:
          type: string
        recordStatus:
          type: string
        fileId:
          type: string
        fileName:
          type: string
        sizeBytes:
          type: integer
        fileType:
          type: string

    VerificationResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VerificationItem"
        total:
          type: integer

    StubCoverageByType:
      type: object
      required: [recordType, fileCount, stubCount, missingStubCount, coveragePercent, totalSizeBytes]
      properties:
        recordType:
          type: string
        fileCount:
          type: integer
        stubCount:
          type: integer
        missingStubCount:
          type: integer
        coveragePercent:
          type: number
        totalSizeBytes:
          type: integer

    CompletionStateRow:
      type: object
      required: [recordType, manuallyDone, notes]
      properties:
        recordType:
          type: string
        manuallyDone:
          type: boolean
        notes:
          type: string

    CompletionStatePatch:
      type: object
      properties:
        manuallyDone:
          type: boolean
        notes:
          type: string

    NetsuiteStatus:
      type: object
      required: [configured, accountId]
      properties:
        configured:
          type: boolean
        accountId:
          type: ["string", "null"]
        message:
          type: ["string", "null"]

    RefreshStubInput:
      type: object
      required: [recordType]
      properties:
        recordType:
          type: string

    RefreshStubResult:
      type: object
      required: [success, updated, message]
      properties:
        success:
          type: boolean
        updated:
          type: integer
        message:
          type: string