# lib/api-zod/src/generated/types/completionStatePatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NetSuite Record-File Viewer API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletionStatePatch {
  manuallyDone?: boolean;
  notes?: string;
}