# lib/api-zod/src/generated/types/completionStateRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NetSuite Record-File Viewer API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletionStateRow {
  recordType: string;
  manuallyDone: boolean;
  notes: string;
}